Move some from 4 to 5
